backend/requirements.txt
Flask==1.1.2
Flask_Cors==3.0.10
Flask_SQLAlchemy==2.5.1
protobuf==3.20.1
PyJWT==2.3.0
python-dotenv==0.20.0
python_bcrypt==0.3.2
SQLAlchemy==1.4.22
Jinja2==3.0.3
itsdangerous==2.0.1
werkzeug==2.0.3